libs.py
import discord
from discord.ext import commands
from discord import app_commands
from discord.ext.commands import CommandError
from PIL import Image
from utils.env import get_env
import os
import random
from discord.errors import DiscordException
from logs.logger import logger