requirements.txt
aiohttp==3.7.4.post0
anyio==3.7.1
asgiref==3.8.1
async-timeout==3.0.1
attrs==23.2.0
chardet==4.0.0
discord==2.3.2
discord.py==2.3.2
Django==5.0.4
idna==3.7
image==1.5.33
logging==0.4.9.6
multidict==6.0.5
pillow==10.3.0
python-dotenv==0.19.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
watchgod==0.8.2
yarl==1.9.4
